+++ cider-connection.el
@@ -798,7 +798,10 @@ session."
                    (:auto (if (eq cur-type 'multi)
                               '(clj cljs)
                             cur-type))))
-           (repls (cider-repls type 'ensure)))
+           (ensure (cl-case which
+                     (:auto nil)
+                     (t 'ensure)))
+           (repls (cider-repls type ensure)))
       (mapcar function repls))))
 
 ;; REPLs double as connections in CIDER, so it's useful to be able to refer to