+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Changes
 
 * [#2496](https://github.com/clojure-emacs/cider/issues/2496): Replace CIDER's pprint implementation with nREPL 0.5's built-in pprint support.
-* [#2558](https://github.com/clojure-emacs/cider/pull/2558): Load clj & cljc files on `cider-load-all-files`.
+* [#2558](https://github.com/clojure-emacs/cider/pull/2558): Load clj, cljc, & cljs (if cljs repl available) files on `cider-load-all-files` (`C-c C-M-l`). Previously, this only loaded clj files.
 
 ## 0.19.0 (2019-01-01)